fix(ecommerce): move getTypes function into Module component
we need it to be available in the client so it can run on the route change
